+++ src/pydvl/value/least_core/montecarlo.py
@@ -9,7 +9,10 @@ from pydvl.utils import Utility, maybe_progress
 from pydvl.utils.config import ParallelConfig
 from pydvl.utils.numeric import random_powerset
 from pydvl.utils.parallel import MapReduceJob
-from pydvl.value.least_core._common import _solve_linear_program
+from pydvl.value.least_core._common import (
+    _solve_egalitarian_least_core_quadratic_program,
+    _solve_least_core_linear_program,
+)
 from pydvl.value.results import ValuationResult, ValuationStatus
 
 logger = logging.getLogger(__name__)
@@ -44,8 +47,7 @@ def _montecarlo_least_core(
         max_subsets=n_iterations,
     )
 
-    A_ub = np.zeros((n_iterations, n + 1))
-    A_ub[:, -1] = -1
+    A_lb = np.zeros((n_iterations, n))
 
     for i, subset in enumerate(
         maybe_progress(
@@ -55,22 +57,22 @@ def _montecarlo_least_core(
             position=job_id,
         )
     ):
-        indices = np.zeros(n + 1, dtype=bool)
+        indices = np.zeros(n, dtype=bool)
         indices[list(subset)] = True
-        A_ub[i, indices] = -1
+        A_lb[i, indices] = 1
         utility_values[i] = u(subset)
 
-    return utility_values, A_ub
+    return utility_values, A_lb
 
 
 def _reduce_func(
     results: Iterable[Tuple[NDArray[np.float_], NDArray[np.float_]]]
 ) -> Tuple[NDArray[np.float_], NDArray[np.float_]]:
     """Combines the results from different parallel runs of the `_montecarlo_least_core` function"""
-    utility_values_list, A_ub_list = zip(*results)
+    utility_values_list, A_lb_list = zip(*results)
     utility_values = np.concatenate(utility_values_list)
-    A_ub = np.concatenate(A_ub_list)
-    return utility_values, A_ub
+    A_lb = np.concatenate(A_lb_list)
+    return utility_values, A_lb
 
 
 def montecarlo_least_core(
@@ -79,7 +81,7 @@ def montecarlo_least_core(
     n_jobs: int = 1,
     config: ParallelConfig = ParallelConfig(),
     *,
-    epsilon: float = 0.01,
+    epsilon: float = 0.0,
     options: Optional[dict] = None,
     progress: bool = False,
 ) -> ValuationResult:
@@ -89,7 +91,7 @@ def montecarlo_least_core(
     \begin{array}{lll}
     \text{minimize} & \displaystyle{e} & \\
     \text{subject to} & \displaystyle\sum_{i\in N} x_{i} = v(N) & \\
-    & \displaystyle\sum_{i\in S} x_{i} + e \geq v(S) & ,
+    & \displaystyle\sum_{i\in S} x_{i} + e + \epsilon \geq v(S) & ,
     \forall S \in \{S_1, S_2, \dots, S_m \overset{\mathrm{iid}}{\sim} U(2^N) \}
     \end{array}
     $$
@@ -98,7 +100,8 @@ def montecarlo_least_core(
 
     * $U(2^N)$ is the uniform distribution over the powerset of $N$.
     * $m$ is the number of subsets that will be sampled and whose utility will be computed
-      and used to compute the Least Core values.
+      and used to compute the data values.
+    * $\epsilon \ge 0$ is an optional relaxation value.
 
     :param u: Utility object with model, data, and scoring function
     :param n_iterations: total number of iterations to use
@@ -106,11 +109,11 @@ def montecarlo_least_core(
     :param config: Object configuring parallel computation, with cluster
         address, number of cpus, etc.
     :param epsilon: Relaxation value by which the subset utility is decreased.
-    :param options: LP Solver options. Refer to `SciPy's documentation
-        <https://docs.scipy.org/doc/scipy/reference/optimize.linprog-highs.html>`_
-        for more information
-    :param progress: Whether to display a progress bar
-    :return: Object with the data values.
+    :param options: Keyword arguments that will be used to select a solver
+        and to configure it. Refer to the following page for all possible options:
+        https://www.cvxpy.org/tutorial/advanced/index.html#setting-solver-options
+    :param progress: If True, shows a tqdm progress bar
+    :return: Object with the data values and the least core value.
     """
     n = len(u.data)
 
@@ -144,7 +147,7 @@ def montecarlo_least_core(
         config=config,
     )
     logger.debug("Calling MapReduceJob instance")
-    utility_values, A_ub = map_reduce_job()
+    utility_values, A_lb = map_reduce_job()
 
     if np.any(np.isnan(utility_values)):
         warnings.warn(
@@ -153,26 +156,44 @@ def montecarlo_least_core(
         )
 
     logger.debug("Building vectors and matrices for linear programming problem")
-    c = np.zeros(n + 1)
-    c[-1] = 1
-    b_ub = -(utility_values - epsilon)
-    # We explicitly add the utility for the empty set
-    A_ub = np.concatenate([np.zeros((1, n + 1)), A_ub], axis=0)
-    A_ub[0, -1] = -1
-    b_ub = np.concatenate([[u([])], b_ub])
-    A_eq = np.ones((1, n + 1))
-    A_eq[:, -1] = 0
+    b_lb = utility_values
+    A_eq = np.ones((1, n))
     # We explicitly add the utility value for the entire dataset
     b_eq = np.array([u(u.data.indices)])
 
-    logger.debug("Removing possible duplicate values in upper bound array")
-    A_ub, unique_indices = np.unique(A_ub, return_index=True, axis=0)
-    b_ub = b_ub[unique_indices]
+    logger.debug("Removing possible duplicate values in lower bound array")
+    A_lb, unique_indices = np.unique(A_lb, return_index=True, axis=0)
+    b_lb = b_lb[unique_indices]
 
-    logger.debug(f"{unique_indices=}")
+    _, subsidy = _solve_least_core_linear_program(
+        A_eq=A_eq, b_eq=b_eq, A_lb=A_lb, b_lb=b_lb, epsilon=epsilon, **options
+    )
 
-    values = _solve_linear_program(
-        c, A_eq, b_eq, A_ub, b_ub, bounds=[(None, None)] * n + [(0.0, None)], **options
+    values: Optional[NDArray[np.float_]]
+
+    if subsidy is None:
+        logger.debug("No values were found")
+        status = ValuationStatus.Failed
+        values = np.empty(n)
+        values[:] = np.nan
+        subsidy = np.nan
+
+        return ValuationResult(
+            algorithm="montecarlo_least_core",
+            status=status,
+            values=values,
+            subsidy=subsidy,
+            stderr=None,
+            data_names=u.data.data_names,
+        )
+
+    values = _solve_egalitarian_least_core_quadratic_program(
+        subsidy,
+        A_eq=A_eq,
+        b_eq=b_eq,
+        A_lb=A_lb,
+        b_lb=b_lb,
+        **options,
     )
 
     if values is None:
@@ -180,18 +201,15 @@ def montecarlo_least_core(
         status = ValuationStatus.Failed
         values = np.empty(n)
         values[:] = np.nan
+        subsidy = np.nan
     else:
         status = ValuationStatus.Converged
 
-    # The last entry represents the least core value 'e'
-    least_core_value = values[-1].item()
-    values = values[:-1]
-
     return ValuationResult(
-        algorithm="exact_least_core",
+        algorithm="montecarlo_least_core",
         status=status,
         values=values,
+        subsidy=subsidy,
         stderr=None,
         data_names=u.data.data_names,
-        least_core_value=least_core_value,
     )