+++ src/pydvl/value/least_core/naive.py
@@ -3,9 +3,13 @@ import warnings
 from typing import Optional
 
 import numpy as np
+from numpy.typing import NDArray
 
 from pydvl.utils import Utility, maybe_progress, powerset
-from pydvl.value.least_core._common import _solve_linear_program
+from pydvl.value.least_core._common import (
+    _solve_egalitarian_least_core_quadratic_program,
+    _solve_least_core_linear_program,
+)
 from pydvl.value.results import ValuationResult, ValuationStatus
 
 __all__ = ["exact_least_core"]
@@ -38,11 +42,12 @@ def exact_least_core(
     Where $N = \{1, 2, \dots, n\}$ are the training set's indices.
 
     :param u: Utility object with model, data, and scoring function
-    :param options: LP Solver options. Refer to `SciPy's documentation
-        <https://docs.scipy.org/doc/scipy/reference/optimize.linprog-highs.html>`_
-        for more information
-    :param progress: Whether to display a progress bar
-    :return: Object with the data values.
+    :param options: Keyword arguments that will be used to select a solver
+        and to configure it. Refer to the following page for all possible options:
+        https://www.cvxpy.org/tutorial/advanced/index.html#setting-solver-options
+    :param progress: If True, shows a tqdm progress bar
+
+    :return: Object with the data values and the least core value.
     """
     n = len(u.data)
 
@@ -56,12 +61,8 @@ def exact_least_core(
     powerset_size = 2**n
 
     logger.debug("Building vectors and matrices for linear programming problem")
-    c = np.zeros(n + 1)
-    c[-1] = 1
-    A_eq = np.ones((1, n + 1))
-    A_eq[:, -1] = 0
-    A_ub = np.zeros((powerset_size, n + 1))
-    A_ub[:, -1] = -1
+    A_eq = np.ones((1, n))
+    A_lb = np.zeros((powerset_size, n))
 
     logger.debug("Iterating over all subsets")
     utility_values = np.zeros(powerset_size)
@@ -73,21 +74,42 @@ def exact_least_core(
             position=0,
         )
     ):
-        indices = np.zeros(n + 1, dtype=bool)
+        indices = np.zeros(n, dtype=bool)
         indices[list(subset)] = True
-        A_ub[i, indices] = -1
+        A_lb[i, indices] = 1
         utility_values[i] = u(subset)
 
-    b_ub = -utility_values
+    b_lb = utility_values
     b_eq = utility_values[-1:]
 
-    values = _solve_linear_program(
-        c,
-        A_eq,
-        b_eq,
-        A_ub,
-        b_ub,
-        bounds=[(None, None)] * n + [(0.0, None)],
+    _, subsidy = _solve_least_core_linear_program(
+        A_eq=A_eq, b_eq=b_eq, A_lb=A_lb, b_lb=b_lb, **options
+    )
+
+    values: Optional[NDArray[np.float_]]
+
+    if subsidy is None:
+        logger.debug("No values were found")
+        status = ValuationStatus.Failed
+        values = np.empty(n)
+        values[:] = np.nan
+        subsidy = np.nan
+
+        return ValuationResult(
+            algorithm="exact_least_core",
+            status=status,
+            values=values,
+            subsidy=subsidy,
+            stderr=None,
+            data_names=u.data.data_names,
+        )
+
+    values = _solve_egalitarian_least_core_quadratic_program(
+        subsidy,
+        A_eq=A_eq,
+        b_eq=b_eq,
+        A_lb=A_lb,
+        b_lb=b_lb,
         **options,
     )
 
@@ -96,18 +118,15 @@ def exact_least_core(
         status = ValuationStatus.Failed
         values = np.empty(n)
         values[:] = np.nan
+        subsidy = np.nan
     else:
         status = ValuationStatus.Converged
-
-    # The last entry represents the least core value 'e'
-    least_core_value = values[-1].item()
-    values = values[:-1]
 
     return ValuationResult(
         algorithm="exact_least_core",
         status=status,
         values=values,
+        subsidy=subsidy,
         stderr=None,
         data_names=u.data.data_names,
-        least_core_value=least_core_value,
     )